+++ packages/plugin-file-upload/web/src/getAddButtons.tsx
@@ -10,7 +10,7 @@ const handleExternalFileChange =
     if (data instanceof Array) {
       data.forEach((file, index) => {
         setTimeout(() => {
-          const nodeId = editorCommands.insertBlockWithBlankLines?.(
+          const nodeId = editorCommands.insertBlockWithBlankLines(
             FILE_UPLOAD_TYPE,
             {},
             { updateSelection: data.length === index + 1 }
@@ -24,7 +24,7 @@ const handleExternalFileChange =
         });
       });
     } else {
-      const nodeId = editorCommands.insertBlockWithBlankLines?.(FILE_UPLOAD_TYPE, {});
+      const nodeId = editorCommands.insertBlockWithBlankLines(FILE_UPLOAD_TYPE, {});
       updateService.updatePluginData({ data }, nodeId, FILE_UPLOAD_TYPE, filePluginService);
     }
   };
@@ -34,7 +34,7 @@ const handleNativeFileChange =
     files.forEach((file, index) => {
       // prevents rerenders when next file starts uploading
       setTimeout(() => {
-        const nodeId = editorCommands.insertBlockWithBlankLines?.(
+        const nodeId = editorCommands.insertBlockWithBlankLines(
           FILE_UPLOAD_TYPE,
           {},
           { updateSelection: files.length === index + 1 }