+++ packages/tiptap-editor/web/src/components/RichContentAdapter/RichContentAdapter.tsx
@@ -190,7 +190,7 @@ export class RichContentAdapter implements TiptapAdapter {
       insertDecoration: (type, data) => {
         const decorationCommandMap = {
           [RICOS_LINK_TYPE]: data => ({ command: 'setLink', args: { link: data } }),
-          [RICOS_ANCHOR_TYPE]: data => ({ command: 'setAnchor', args: data.anchor }),
+          [RICOS_ANCHOR_TYPE]: data => ({ command: 'setAnchor', args: data }),
           [RICOS_TEXT_COLOR_TYPE]: data => ({ command: 'setColor', args: data.color }),
           [RICOS_TEXT_HIGHLIGHT_TYPE]: data => ({ command: 'setHighlight', args: data.color }),
           [RICOS_MENTION_TYPE]: data => ({ command: 'insertMention', args: data.mention }),
@@ -314,6 +314,9 @@ export class RichContentAdapter implements TiptapAdapter {
           // eslint-disable-next-line @typescript-eslint/ban-ts-comment
           // @ts-ignore
           [RICOS_LINK_TYPE]: () => this.tiptapEditor.commands.unsetLink(),
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          [RICOS_ANCHOR_TYPE]: () => this.tiptapEditor.commands.unsetAnchor(),
           // eslint-disable-next-line @typescript-eslint/ban-ts-comment
           // @ts-ignore
           [RICOS_INDENT_TYPE]: () => this.tiptapEditor.commands.outdent(),